+++ e2e/setup-jest.js
@@ -1,7 +1,7 @@
 const path = require( 'path' );
 const childProcess = require( 'child_process' );
 
-childProcess.execSync( 'yarn build:silent', {
+childProcess.execSync( 'yarn build', {
     cwd: path.resolve( __dirname, '..' )
 } );